Move ComposerPackageUsageAnalyzer to Composer
